Remove incorrect license info
Quake is licensed under GPL-2.0, which requires that all derivative
works are also GPL-2.0. Seismon still keeps Cormac O'Brien's
licenses for now since the rendering code very much is his own, but
the QC interpreter is uncontestably a derivative work of the Quake
GPL release. Apart from that, at this point this part of the codebase
contains almost none of his code - it was by far the most incomplete
part of the codebase when it was forked to become Seismon.